# backend/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "DiscoveryShop Marketplace API",
    "version": "1.0.0",
    "description": "Professional API documentation for DiscoveryShop - Full-stack e-commerce marketplace platform with AI-powered discovery, dual-role accounts, real-time chat, and comprehensive admin controls.",
    "contact": {
      "name": "API Support",
      "email": "support@discoveryshop.local"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development Server"
    },
    {
      "url": "https://api.discoveryshop.com",
      "description": "Production Server"
    }
  ],
  "paths": {
    "/api/v1/auth/register": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Register new user",
        "description": "Create a new user account with email, username, and password. Returns JWT token for authentication.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "User email address"
                  },
                  "username": {
                    "type": "string",
                    "description": "Username (3-30 alphanumeric characters)"
                  },
                  "password": {
                    "type": "string",
                    "description": "Password (min 8 characters, 1 uppercase, 1 lowercase, 1 digit)"
                  }
                },
                "required": ["email", "username", "password"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user_id": {"type": "integer"},
                    "email": {"type": "string"},
                    "username": {"type": "string"},
                    "token": {"type": "string"},
                    "created_at": {"type": "string", "format": "date-time"}
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          },
          "409": {
            "description": "Email/username already exists"
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "User login",
        "description": "Authenticate user with email/username and password. Returns JWT token valid for 24 hours.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "description": "Email or username"
                  },
                  "password": {
                    "type": "string",
                    "description": "User password"
                  }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user_id": {"type": "integer"},
                    "email": {"type": "string"},
                    "username": {"type": "string"},
                    "token": {"type": "string"},
                    "expires_in": {"type": "integer"}
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid credentials"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from login endpoint"
      }
    },
...
  },
...
}